refactor(orchestrator): drop quota warning scoring superseded by routing gate
